# Improve performance of `dpnp.matmul` and `dpnp.dot` with `out` keyword