In this PR, implementation of `dpnp.matmul` and `dpnp.dot` is updated to improve the performance when `out` keyword is present and its feature matches the appropriate shape, dtype, sycl_queue, and usm_type that is needed for performing the calculation in the corresponding OneMKL BLAS routines.

Sample timing including this update
```
>>> import dpnp
>>> a = dpnp.ones((21846, 3), device='cpu')
>>> b = dpnp.ones((21846, 3), device='cpu')
>>> c = dpnp.ones((21846, 21846), device='cpu')

>>> %timeit dpnp.matmul(a, b.T, c)
229 ms ± 22 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

Sample timing excluding this update
```
>>> import dpnp
>>> a = dpnp.ones((21846, 3), device='cpu')
>>> b = dpnp.ones((21846, 3), device='cpu')
>>> c = dpnp.ones((21846, 21846), device='cpu')

>>> %timeit dpnp.matmul(a, b.T, c)
798 ms ± 30.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

- [x] Have you provided a meaningful PR description?
- [ ] Have you added a test, reproducer or referred to issue with a reproducer?
- [x] Have you tested your changes locally for CPU and GPU devices?
- [x] Have you made sure that new changes do not introduce compiler warnings?
- [x] Have you checked performance impact of proposed changes?
- [ ] If this PR is a work in progress, are you filing the PR as a draft?